Pickup/Drop
Can pickup with E and Drop with G.

Need to have a "WepSlot" empty gameobject in player capsule for where the object will go after it gets picked up.

Script is called "PickUp" and must be placed on all pickuppable objects.

WepSlot needs to be adjusted so the weapon doesnt block a majority of the FOV.